feat: add support for @koa/router
The `koa-router` module is abandoned and a fork is being maintained
under `@koa/router` instead.

This fork is published as a major bump (v8.0.0) of the original, but
our patch still works fine on that version.

Closes #1343
