+++ .ci/.jenkins_tav.yml
@@ -1,30 +1,31 @@
 TAV:
+  - '@koa/router'
+  - apollo-server-express
+  - bluebird
+  - cassandra-driver
+  - elasticsearch
+  - express
+  - express-graphql
+  - express-queue
+  - fastify
+  - finalhandler
   - generic-pool
+  - got
+  - graphql
+  - handlebars
+  - hapi
+  - ioredis
+  - jade
+  - knex
+  - koa-router
+  - memcached
+  - mimic-response
+  - mongodb-core
   - mysql
   - mysql2
-  - redis
-  - koa-router
-  - handlebars
-  - jade
-  - pug
-  - mongodb-core
-  - finalhandler
-  - ioredis
   - pg
-  - cassandra-driver
-  - tedious
+  - pug
+  - redis
   - restify
-  - fastify
-  - mimic-response
-  - got
-  - bluebird
-  - apollo-server-express
-  - knex
+  - tedious
   - ws
-  - graphql
-  - express-graphql
-  - elasticsearch
-  - hapi
-  - express
-  - express-queue
-  - memcached
