+++ .tav.yml
@@ -201,7 +201,13 @@ koa-router:
   node: '>=6.0.0'
   peerDependencies: koa@2
   versions: '>=5.2.0 <8'
-  commands: node test/instrumentation/modules/koa-router/index.js
+  commands: node test/instrumentation/modules/koa-router/old-name.js
+'@koa/router':
+  node: '>=8.0.0'
+  peerDependencies: koa@2
+  versions: '>=8 <9'
+  commands: node test/instrumentation/modules/koa-router/new-name.js
+
 elasticsearch:
   versions: '>=8.0.0'
   commands: node test/instrumentation/modules/elasticsearch.js
