+++ docs/supported-technologies.asciidoc
@@ -47,7 +47,7 @@ These are the frameworks that we officially support:
 |<<express,Express>> |^4.0.0 |
 |<<hapi,hapi>> |>=9.0.0 <19.0.0 |
 |<<hapi,@hapi/hapi>> |>=17.9.0 <19.0.0 |
-|<<koa,Koa>> via koa-router |>=5.2.0 <8.0.0 |Koa doesn't have a built in router,
+|<<koa,Koa>> via koa-router or @koa/router |>=5.2.0 <9.0.0 |Koa doesn't have a built in router,
 so we can't support Koa directly since we rely on router information for full support.
 We currently support the most popular Koa router called https://github.com/alexmingoia/koa-router[koa-router]
 |<<restify,Restify>> |>=5.2.0 |
