+++ lib/instrumentation/index.js
@@ -31,7 +31,7 @@ var MODULES = [
   'jade',
   'knex',
   'koa',
-  'koa-router',
+  ['koa-router', '@koa/router'],
   'memcached',
   'mimic-response',
   'mongodb-core',
