+++ lib/instrumentation/modules/koa-router.js
@@ -6,7 +6,7 @@ var shimmer = require('../shimmer')
 
 module.exports = function (Router, agent, { version, enabled }) {
   if (!enabled) return Router
-  if (!semver.satisfies(version, '>=5.2.0 <8')) {
+  if (!semver.satisfies(version, '>=5.2.0 <9')) {
     agent.logger.debug('koa-router version %s not supported - aborting...', version)
     return Router
   }
