+++ package.json
@@ -110,6 +110,7 @@
     "@commitlint/config-conventional": "^8.0.0",
     "@commitlint/travis-cli": "^8.0.0",
     "@hapi/hapi": "^18.2.0",
+    "@koa/router": "^8.0.1",
     "@types/node": "^12.0.8",
     "apollo-server-express": "^2.6.3",
     "aws-sdk": "^2.477.0",
