+++ test/instrumentation/modules/koa-router/new-name.js
@@ -1,0 +1,3 @@
+'use strict'
+
+require('./shared')('@koa/router')
